Revert temporary file rename
