+++ libs/services/designer-client-services/src/lib/base/operationmanifest.ts
