+++ libs/services/designer-client-services/src/lib/standard/connection.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign */
-import { azureFunctionConnectorId } from '../base/tempoperationmanifest';
+import { azureFunctionConnectorId } from '../base/operationmanifest';
 import type { HttpResponse } from '../common/exceptions/service';
 import type {
   ConnectionCreationInfo,
